Improve code

+++ app/src/main/java/com/dimonkiv/savingstracker/feature/account/domain/model/Account.kt
@@ -4,7 +4,7 @@ data class Account(
     val id: Long = 0,
     val typeId: Long = 0,
     val title: String = "",
-    val balance: Int = 0,
+    val balance: Long = 0L,
     val color: String,
     val icon: String
 )
